add validation tests in Marketplace, lint check, remove ddoValidation
