+++ package.json
@@ -22,7 +22,6 @@
     "prepublishOnly": "npm run build",
     "test": "npm run lint && npm run test:unit:cover && npm run test:integration:cover",
     "test:unit": "mocha --config=test/unit/.mocharc.json --node-env=test --exit 'test/unit/**/*.test.ts'",
-    "test:ddoValidation": "mocha --config=test/ddoValidation/.mocharc.json --node-env=test --exit 'test/ddoValidation/**/*.test.ts'",
     "test:unit:cover": "nyc --report-dir coverage/unit npm run test:unit",
     "test:integration": "mocha --config=test/integration/.mocharc.json --node-env=test --exit 'test/integration/**/*.test.ts'",
     "test:integration:cover": "nyc --report-dir coverage/integration --no-clean npm run test:integration"
