+++ src/ocean/Assets.ts
@@ -82,21 +82,14 @@ export class Assets extends Instantiable {
       )
       return null
     }
-    
     this.logger.log('Creating asset')
     return new SubscribablePromise(async (observer) => {
-     
       const isValidDDO = await this.ocean.metadataCache.validateMetadata(metadata)
-    //console.log(isValidDDO)
-      
-     if (isValidDDO !== true) {
-     this.logger.error(
-       `Passed Metadata is not valid. Aborting publishing.`
-     )
-     return null;
-     }
-
-    if (services.length === 0) {
+      if (isValidDDO !== true) {
+        this.logger.error(`Passed Metadata is not valid. Aborting publishing.`)
+        return null
+      }
+      if (services.length === 0) {
         this.logger.log('You have no services. Are you sure about this?')
       }
       const { datatokens } = this.ocean
@@ -191,8 +184,6 @@ export class Assets extends Instantiable {
             index: indexCount++
           })) as Service[]
       })
-     
-    
 
       await ddo.addProof(this.ocean, publisher.getId())
       ddo.dataTokenInfo = {
@@ -203,8 +194,6 @@ export class Assets extends Instantiable {
       }
       this.logger.log('Storing DDO')
       observer.next(CreateProgressStep.StoringDdo)
-   
-   
 
       const storeTx = await this.ocean.onChainMetadata.publish(
         ddo.id,
@@ -213,11 +202,10 @@ export class Assets extends Instantiable {
       )
       this.logger.log('DDO stored ' + ddo.id)
       observer.next(CreateProgressStep.DdoStored)
-        
+
       if (storeTx) return ddo
       else return null
-    }) 
-    
+    })
   }
 
   /**
@@ -270,9 +258,7 @@ export class Assets extends Instantiable {
    * @return {Promise<DDO>} the new DDO
    */
   public async editMetadata(ddo: DDO, newMetadata: EditableMetadata): Promise<DDO> {
-   
     if (!ddo) return null
-    
 
     for (let i = 0; i < ddo.service.length; i++) {
       if (ddo.service[i].type !== 'metadata') continue
@@ -283,9 +269,8 @@ export class Assets extends Instantiable {
         ddo.service[i].attributes.additionalInformation.description =
           newMetadata.description
       if (newMetadata.links)
-        ddo.service[i].attributes.additionalInformation.links = newMetadata.links       
+        ddo.service[i].attributes.additionalInformation.links = newMetadata.links
     }
-    
 
     return ddo
   }
@@ -301,14 +286,10 @@ export class Assets extends Instantiable {
     consumerAccount: string
   ): Promise<TransactionReceipt> {
     const isValidDDO = await this.ocean.metadataCache.validateMetadataRemote(ddo)
-    //console.log(isValidDDO)
-      
-     if (isValidDDO !== true) {
-     this.logger.error(
-       `Passed Metadata is not valid. Aborting publishing.`
-     )
-     return null;
-     }
+    if (isValidDDO !== true) {
+      this.logger.error(`Passed Metadata is not valid. Aborting publishing.`)
+      return null
+    }
     return await this.ocean.onChainMetadata.update(ddo.id, ddo, consumerAccount)
   }
